% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_imports.R
\docType{package}
\name{unodctools-package}
\alias{unodctools}
\alias{unodctools-package}
\title{unodctools: Tools for UNODC SDMX Data Checks and Excel Sensitivity Labeling}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Markus_Schwabe Christian_Halaby \email{christian.halaby@un.org}

}
\keyword{internal}
